49 finetuning program

+++ src/main/java/com/spring/SpeedAuction/security/Services/AuctionTypeCarService.java
@@ -61,8 +61,14 @@ public class AuctionTypeCarService {
     }
 
     public String deleteAuctionTypeCar(String id) {
-        auctionTypeCarRepository.deleteById(id);
-        return "AuctionTypeCar deleted";
+        AuctionTypeCar auctionTypeCar = auctionTypeCarRepository.findById(id).orElse(null);
+        if (auctionTypeCar != null) {
+            auctionTypeCarRepository.deleteById(id);
+            return "auctionTypeCar deleted";
+        }
+        else {
+            return "auctionTypeCar id doesnt exist";
+        }
     }
     public AuctionTypeCar createAuctiontypecar (CarDTO auctionTypeCarDTO) {
 
@@ -72,7 +78,6 @@ public class AuctionTypeCarService {
 
 
         AuctionTypeCar newAuctionTypeCar = new AuctionTypeCar();
-        //newAuctionTypeCar.setId(auctionTypeCarDTO.getId());
         newAuctionTypeCar.setAuction(auction);
         newAuctionTypeCar.setCarModel(auctionTypeCarDTO.getCarModel());
         newAuctionTypeCar.setBrand(auctionTypeCarDTO.getBrand());
@@ -87,7 +92,7 @@ public class AuctionTypeCarService {
         AuctionModels existingAuction = auctionRepository.findById(newAuctionTypeCar.getAuction().getId()).orElseThrow(() -> new IllegalArgumentException("auction does not exist"));
         existingAuction.setActive(true);
         existingAuction.setSeller(existingAuction.getSeller());
-        existingAuction.setStartingBid(existingAuction.getStartingBid());
+        existingAuction.setStartingPrice(existingAuction.getStartingPrice());
         existingAuction.setCreated_at(existingAuction.getCreated_at());
         existingAuction.setEndOfAuction(existingAuction.getEndOfAuction());
         existingAuction.setUpdated_at(existingAuction.getUpdated_at());
@@ -105,7 +110,6 @@ public class AuctionTypeCarService {
     }
     private CarDTO convertToDTO(AuctionTypeCar auctionTypeCar) {
         CarDTO carDTOResponse = new CarDTO();
-        //carDTOResponse.setId(auctionTypeCar.getId());
         carDTOResponse.setAuctionId(auctionTypeCar.getAuction().getId());
         carDTOResponse.setBrand(auctionTypeCar.getBrand());
         carDTOResponse.setCarModel(auctionTypeCar.getCarModel());
